% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_DE_MRF.R
\name{get_DE_MRF}
\alias{get_DE_MRF}
\title{Run the MRF model to estimate posterior probabilities of differential expression for each gene across each cell type}
\usage{
get_DE_MRF(
  data,
  g_g,
  c_c,
  nulltype = 1,
  df = 15,
  iterEM = 200,
  iterGibbsPost = 20000,
  brPost = 10000
)
}
\arguments{
\item{g_g}{Gene to gene network matrix}

\item{c_c}{Cell to cell dependency matrix}

\item{nulltype}{Type of null hypothesis assumed in estimating f0, see locfdr package.Default is the MLE (nulltype=1)}

\item{df}{Degrees of freedom for fitting the estimated density, see locfdr package. Default df=15}

\item{iterEM}{Max number of iterations for the EM algorithm. Default=200}

\item{iterGibbsPost}{Number of Gibbs posterior samples. Default=20,000}

\item{brPost}{Number of burn-in for the posterior samples. Default=10,000}

\item{zz}{Summary statistics matrix, rows are genes, columns are cell types}
}
\value{
The estimated model parameters and the posterior probabilities of differential expression
\item{postDE}{Posterior probabilities of differential expression. A 2-dimensional array: (num of genes)*(num of cell types)}
\item{paraMRF}{Estimated model parameters}
\item{paraMRFTrace}{Trace of the estimated model parameters in the EM algorithm}
}
\description{
Run the MRF model to estimate posterior probabilities of differential expression for each gene across each cell type
}
