% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_DE_NOMRF.R
\name{get_DE_noMRF}
\alias{get_DE_noMRF}
\title{Get simulation results for three types of tests}
\usage{
get_DE_noMRF(data, test_type, alpha = 0.01)
}
\arguments{
\item{data}{A list that contains
A 4-dimensional expression array: (2 groups)\emph{(num of replicates)}(num of genes)*(num of cell types)
and other information such as number of genes, number of DE genes, etc.}

\item{test_type}{Choose from "ttest" and "hurdle"}

\item{alpha}{significance level. Default=0.01}
}
\value{
Returns simulation results
}
\description{
Get simulation results for three types of tests
}
