% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_DE_RESULTS.R
\name{get_DE_results}
\alias{get_DE_results}
\title{Get sensitivity, specificity and FDR for the obtained simulation results.}
\usage{
get_DE_results(MRF_Results, alpha = 0.05)
}
\arguments{
\item{alpha}{Significance level. Default = 0.05}

\item{results}{A list that contains MRF results}
}
\value{
Returns a binary matrix that corresponds to cell-type specific DE results.
}
\description{
Get sensitivity, specificity and FDR for the obtained simulation results.
}
