% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_SENS_SPEC_FDR.R
\name{get_sens_spec_fdr}
\alias{get_sens_spec_fdr}
\title{Get sensitivity, specificity and FDR for the obtained simulation results.}
\usage{
get_sens_spec_fdr(data, test_type, results, alpha = 0.05)
}
\arguments{
\item{data}{A list that contains
A 4-dimensional expression array: (2 groups)\emph{(num of replicates)}(num of genes)*(num of cell types)
and other information such as number of genes, number of DE genes, etc.}

\item{test_type}{Choose from "ttest", "hurdle" and "MRF"}

\item{results}{Simulation results. For "ttest" and "hurdle", a matrix of p_values. For "MRF". See getMRFDE().}

\item{alpha}{Significance level. Default=0.01}
}
\value{
Returns sensitivity, specificity, and fdr
}
\description{
Get sensitivity, specificity and FDR for the obtained simulation results.
}
