% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_SIM_DATA.R
\name{get_sim_data}
\alias{get_sim_data}
\title{Get simulated data}
\usage{
get_sim_data(gg_info, gamma_alpha, gamma_beta, tau, random.seed)
}
\arguments{
\item{gg_info}{A list that contains gene to gene network, DE genes and non-DE genes}

\item{gamma_alpha}{Shape parameter of Gamma distribution}

\item{gamma_beta}{Scale parameter of Gamma distribution}

\item{tau}{DE effect}

\item{random.seed}{A random seed}
}
\value{
A simulated data set
}
\description{
Get simulated data
}
\examples{
data = get_sim_data(gg_info, gamma_alpha = 2, gamma_beta = 4, tau = 5, random.seed = 2020)
}
