% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GET_Z_SCORES.R
\name{get_z_scores}
\alias{get_z_scores}
\title{Get z scores}
\usage{
get_z_scores(data)
}
\arguments{
\item{data}{A list that contains
A 4-dimensional expression array: (2 groups)\emph{(num of replicates)}(num of genes)*(num of cell types)
and other information such as number of genes, number of DE genes, etc.}
}
\value{
Returns z scores
}
\description{
Get z scores
}
